Add CSS for the credit card
Even though it's the exact same as all the other cards in every way except for like two things. For some reason, I couldn't get the width and height set in "card" to reset so it would just fit to the text in the box. Oh, well. This is already a horrible mess.